moved operations into their own file
